Use a lowercase 'immutable' module name
Webpack complains with 'Immutable':

WARNING in ./~/immutable/dist/Immutable.js
There is another module with an equal name when case is ignored.
This can lead to unexpected behavior when compiling on a filesystem
with other case-semantic.
Rename module if multiple modules are expected or use equal casing if
one module is expected.